+++ src/components/index.ts
@@ -3,7 +3,6 @@ export { default as CachedImage } from "./CachedImage";
 export { default as CartPage } from "./CartPage";
 export { default as ContentPage } from "./ContentPage";
 export { default as HomePage } from "./HomePage";
-export { default as ProductPage } from "./ProductPage";
 export { default as Button, ButtonProps } from "./Button";
 export { default as TextField } from "./TextField";
 export { default as SelectField } from "./SelectField";
@@ -44,3 +43,4 @@ export { Footer } from "./Footer";
 export { MainMenu, MainMenuNavOverlay } from "./MainMenu";
 export { MobileNav } from "./MobileNav";
 export { NavLink } from "./NavLink";
+export { ProductPage } from "./ProductPage";
